fix(mac): unable to paste files in some apps (#27)
In macOS 10.6 and later, the functionality of type declarations has been covered by `writeObjects(_:)`, so there is no need to call declare_types anymore. In previous implementations, calling declare_types caused issues with the clipboard data, resulting in the ability to paste files only in Finder, while pasting in apps like Slack and WeChat was not possible.

Signed-off-by: Kevin Cui <bh@bugs.cc>